Refactor/Migrate season stat helper methods
